Clean up database class
