--- tests/database.json
@@ -1,1 +1,0 @@
-{ "africa": "music-name", "toto": "music-band", "toto IV": "music-album", "tot": "music-playlist" }
